Refactoring and javadocs of polyline updater
